requirements.txt
Pillow>=9.0.0
waveshare-epd>=0.0.1
